requirement.txt
opencv-python==4.5.5.62
numpy==1.19.5
scikit-image==0.17.2
pillow
pyyaml==5.3.1
yacs
tqdm
pandas
lpips